+++ CHANGES.rst
@@ -16,6 +16,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Ship buffered region file with the code.
+
 - Stop forwarding client IP address to data pipeline.
 
 20160112143200